requirements.txt
# Core
groq>=0.13.0
python-dotenv>=1.0.0
pydantic>=2.0.0
opencv-python>=4.9.0

# Storage
chromadb>=0.5.0

# LangChain ecosystem
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0

# UI
streamlit>=1.35.0

# Testing
pytest>=8.0.0